se agrega campo favorita - HU25
