+++ flaskr/vistas/VistaCancion.py
@@ -1,0 +1,66 @@
+from flask import request
+from ..modelos import db, Cancion, CancionSchema, Album
+from flask_restful import Resource
+
+
+cancion_schema = CancionSchema()
+
+class VistaCanciones(Resource):
+
+    def post(self):
+        nueva_cancion = Cancion(titulo=request.json["titulo"], minutos=request.json["minutos"], 
+                        segundos=request.json["segundos"], interprete=request.json["interprete"], favorita=request.json["favorita"])
+        db.session.add(nueva_cancion)
+        db.session.commit()
+        return cancion_schema.dump(nueva_cancion)
+
+    def get(self):
+        return [cancion_schema.dump(ca) for ca in Cancion.query.all()]
+        
+
+class VistaCancion(Resource):
+
+    def get(self, id_cancion):
+        return cancion_schema.dump(Cancion.query.get_or_404(id_cancion))
+
+    def put(self, id_cancion):
+        cancion = Cancion.query.get_or_404(id_cancion)
+        cancion.titulo = request.json.get("titulo",cancion.titulo)
+        cancion.minutos = request.json.get("minutos",cancion.minutos)
+        cancion.segundos = request.json.get("segundos",cancion.segundos)
+        cancion.interprete = request.json.get("interprete",cancion.interprete)
+        ## @farojasp1 se adiciona campo para marca favorita - HU25
+        cancion.favorita =request.json.get("favorita",cancion.favorita)
+        db.session.commit()
+        return cancion_schema.dump(cancion)
+
+    def delete(self, id_cancion):
+        cancion = Cancion.query.get_or_404(id_cancion)
+        db.session.delete(cancion)
+        db.session.commit()
+        return '',204
+
+
+class VistaCancionesAlbum(Resource):
+
+    def post(self, id_album):
+        album = Album.query.get_or_404(id_album)
+        
+        if "id_cancion" in request.json.keys():
+            
+            nueva_cancion = Cancion.query.get(request.json["id_cancion"])
+            if nueva_cancion is not None:
+                album.canciones.append(nueva_cancion)
+                db.session.commit()
+            else:
+                return 'Canción errónea',404
+        else: 
+            nueva_cancion = Cancion(titulo=request.json["titulo"], minutos=request.json["minutos"], segundos=request.json["segundos"], interprete=request.json["interprete"])
+            album.canciones.append(nueva_cancion)
+        db.session.commit()
+        return cancion_schema.dump(nueva_cancion)
+       
+    def get(self, id_album):
+        album = Album.query.get_or_404(id_album)
+        return [cancion_schema.dump(ca) for ca in album.canciones]
+
